+++ app/features/charts/charts.stories.tsx
@@ -1,9 +1,11 @@
-import RmetricBalanceDistributionChart from "./top-rmetric-holders";
+import { fakeRMetricDistributionData } from "~/utils/fakes";
+import RmetricBalanceDistributionChart from "./rmetric-balance-distribution";
 
 export function RMetricBalanceDistributionChartExample() {
+  const data = fakeRMetricDistributionData();
   return (
     <div className="h-60">
-      <RmetricBalanceDistributionChart userBalance={100} />;
+      <RmetricBalanceDistributionChart data={data} userBalance={522} />;
     </div>
   );
 }
